+++ apps/designer/app/designer/features/sidebar-left/index.tsx
@@ -1,1 +1,2 @@
 export { SidebarLeft } from "./sidebar-left";
+export { Navigator } from "./navigator";
