+++ apps/designer/app/designer/features/sidebar-left/panels/navigator.tsx
@@ -1,13 +1,22 @@
 import { type Publish } from "@webstudio-is/react-sdk";
 import { ListNestedIcon } from "@webstudio-is/icons";
-import { Navigator } from "~/designer/features/navigator";
+import { Navigator } from "../navigator";
+import { TabName } from "../types";
 
 type TabContentProps = {
   publish: Publish;
+  onSetActiveTab: (tabName: TabName) => void;
 };
 
-export const TabContent = ({ publish }: TabContentProps) => {
-  return <Navigator publish={publish} />;
+export const TabContent = ({ publish, onSetActiveTab }: TabContentProps) => {
+  return (
+    <Navigator
+      publish={publish}
+      onClose={() => {
+        onSetActiveTab("none");
+      }}
+    />
+  );
 };
 
 export const icon = <ListNestedIcon />;
