+++ packages/design-system/src/index.ts
@@ -61,3 +61,4 @@ export * from "./components/combobox";
 export * from "./components/checkbox";
 export * from "./components/avatar";
 export * from "./components/primitives/dnd";
+export * from "./components/separator";
