+++ packages/design-system/src/stitches.config.ts
@@ -209,50 +209,50 @@ export const {
       sans: "Inter, -apple-system, system-ui, sans-serif",
       mono: "Söhne Mono, menlo, monospace",
     },
+    // space: {
+    //   1: "5px",
+    //   2: "10px",
+    //   3: "15px",
+    //   4: "20px",
+    //   5: "25px",
+    //   6: "35px",
+    //   7: "45px",
+    //   8: "65px",
+    //   9: "80px",
+    // },
+    // sizes: {
+    //   1: "5px",
+    //   2: "10px",
+    //   3: "15px",
+    //   4: "20px",
+    //   5: "25px",
+    //   6: "35px",
+    //   7: "45px",
+    //   8: "65px",
+    //   9: "80px",
+    // },
     space: {
-      1: "5px",
-      2: "10px",
-      3: "15px",
+      1: "4px",
+      2: "8px",
+      3: "16px",
       4: "20px",
-      5: "25px",
-      6: "35px",
-      7: "45px",
-      8: "65px",
+      5: "24px",
+      6: "32px",
+      7: "48px",
+      8: "64px",
       9: "80px",
     },
     sizes: {
-      1: "5px",
-      2: "10px",
-      3: "15px",
+      1: "4px",
+      2: "8px",
+      3: "16px",
       4: "20px",
-      5: "25px",
-      6: "35px",
-      7: "45px",
-      8: "65px",
+      5: "24px",
+      6: "32px",
+      7: "48px",
+      8: "64px",
       9: "80px",
     },
-    // space: {
-    //   1: '4px',
-    //   2: '8px',
-    //   3: '16px',
-    //   4: '20px',
-    //   5: '24px',
-    //   6: '32px',
-    //   7: '48px',
-    //   8: '64px',
-    //   9: '80px',
-    // },
-    // sizes: {
-    //   1: '4px',
-    //   2: '8px',
-    //   3: '16px',
-    //   4: '20px',
-    //   5: '24px',
-    //   6: '32px',
-    //   7: '48px',
-    //   8: '64px',
-    //   9: '80px',
-    // },
     fontSizes: {
       1: "12px",
       2: "13px",
